# Performance optimizations: Redis caching, DB pool, compression (Phase 1 & 2)