--- a/.env.example
+++ b/.env.example
@@ -11,6 +11,7 @@ BASE_URL=http://localhost:3000
 ENABLE_ANALYTICS=false
 ENABLE_CUSTOM_CODES=false
 ENABLE_RATE_LIMITING=false
+ENABLE_CACHING=false
 
 # Rate Limiting (if enabled)
 RATE_LIMIT_WINDOW_MS=900000
@@ -22,5 +23,15 @@ DB_PORT=5432
 DB_NAME=url_shortener
 DB_USER=postgres
 DB_PASSWORD=postgres
-DB_POOL_MIN=2
-DB_POOL_MAX=10
+DB_POOL_MIN=5
+DB_POOL_MAX=50
+
+# Redis Configuration (for caching, if enabled)
+# Use REDIS_URL for managed services (Railway, Heroku, etc.)
+# Or use individual host/port/password settings
+# REDIS_URL=redis://default:password@host:port
+REDIS_HOST=localhost
+REDIS_PORT=6379
+# REDIS_PASSWORD=
+REDIS_TTL=3600
+REDIS_KEY_PREFIX=urlshort: