--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -57,6 +57,7 @@ jobs:
           ENABLE_ANALYTICS: true
           ENABLE_CUSTOM_CODES: true
           ENABLE_RATE_LIMITING: false
+          ENABLE_CACHING: false
         run: npm test
 
       - name: Generate coverage report
@@ -73,6 +74,7 @@ jobs:
           ENABLE_ANALYTICS: true
           ENABLE_CUSTOM_CODES: true
           ENABLE_RATE_LIMITING: false
+          ENABLE_CACHING: false
         run: npm run test:coverage
 
       - name: Upload coverage reports