--- a/package.json
+++ b/package.json
@@ -10,6 +10,10 @@
     "start": "node dist/index.js",
     "start:staging": "NODE_ENV=staging node --env-file=.env.staging dist/index.js",
     "migrate": "ts-node src/database/migrate.ts",
+    "migrate:prod": "node dist/database/migrate.js",
+    "migrate:staging": "NODE_ENV=staging node --env-file=.env.staging dist/database/migrate.js",
+    "start:migrate": "npm run migrate:prod && npm start",
+    "start:migrate:staging": "npm run migrate:staging && npm run start:staging",
     "test": "jest",
     "test:watch": "jest --watch",
     "test:coverage": "jest --coverage",
@@ -34,13 +38,17 @@
   "license": "ISC",
   "type": "commonjs",
   "dependencies": {
+    "compression": "^1.8.1",
     "dotenv": "^17.2.3",
     "express": "^5.1.0",
+    "ioredis": "^5.8.0",
     "nanoid": "^3.3.11",
     "pg": "^8.16.3"
   },
   "devDependencies": {
+    "@types/compression": "^1.8.1",
     "@types/express": "^5.0.3",
+    "@types/ioredis": "^4.28.10",
     "@types/jest": "^30.0.0",
     "@types/node": "^24.6.2",
     "@types/pg": "^8.15.5",